rename directory from util to utils.
